Make API calls into Flask views, not working yet.
